Add Google Test, fix warnings, enable Firebird, cleanup build artifacts
- Add build/ to .gitignore, remove build artifacts from git tracking
- Set up Google Test via FetchContent (BUILD_TESTING option)
- Add 28 tests: 7 for CLSID roundtrip, 21 for ibValue type system
- Fix ~66,800 compiler warnings:
  - Add missing override keywords across backend headers
  - Fix member initializer order in 19 files
  - Remove incorrect inline on out-of-line definitions
  - Add CastValue forward declarations to fix C++20 extension warnings
  - Fix infinite recursion in typeconv.h overload resolution
  - Fix -Wswitch, -Wpessimizing-move, -Wrange-loop-construct
- Enable Firebird driver on macOS (find Framework paths)
- Fix Firebird ibNumber ambiguous assignment
- Fix clipboard assert on exit (Open before Flush)

+++ src/engine/backend/databaseLayer/firebird/firebirdResultSet.cpp
@@ -298,7 +298,7 @@ double ibDatabaseResultSetFirebird::GetResultDouble(int nField)
 		}
 		else if (nType == SQL_LONG)
 		{
-			dblReturn = *(long*)(pVar->sqldata);
+			dblReturn = static_cast<int>(*(long*)(pVar->sqldata));
 			for (int i = 0; i < -pVar->sqlscale; dblReturn /= 10, i++);
 		}
 		else if (nType == SQL_INT64)
@@ -350,7 +350,7 @@ ibNumber ibDatabaseResultSetFirebird::GetResultNumber(int nField)
 		}
 		else if (nType == SQL_LONG)
 		{
-			dblReturn = *(long*)(pVar->sqldata);
+			dblReturn = static_cast<int>(*(long*)(pVar->sqldata));
 			for (int i = 0; i < -pVar->sqlscale; i++) dblReturn /= 10;
 		}
 		else if (nType == SQL_INT64)
